+++ source/main/ipc/index.js
@@ -9,6 +9,7 @@ import getGpuStatus from './get-gpu-status';
 import { handleBugReportRequests } from './bugReportRequestChannel';
 import { handlePaperWalletRequests } from './generatePaperWalletChannel';
 import { openExternalUrlChannel } from './open-external-url';
+import { openLocalDirectoryChannel } from './open-local-directory';
 import { setupParseRedemptionCodeHandler } from './parse-redemption-code';
 
 export default (window: BrowserWindow) => {
@@ -22,5 +23,7 @@ export default (window: BrowserWindow) => {
   handlePaperWalletRequests();
   // eslint-disable-next-line no-unused-expressions
   openExternalUrlChannel;
+  // eslint-disable-next-line no-unused-expressions
+  openLocalDirectoryChannel;
   setupParseRedemptionCodeHandler();
 };