+++ storybook/stories/Staking.stories.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, date, number } from '@storybook/addon-knobs';
+import { withKnobs, date, number, radios } from '@storybook/addon-knobs';
 import { linkTo } from '@storybook/addon-links';
 import { action } from '@storybook/addon-actions';
 import StoryLayout from './support/StoryLayout';
@@ -15,6 +15,7 @@ import StakingCountdown from '../../source/renderer/app/components/staking/count
 import StakingInfo from '../../source/renderer/app/components/staking/info/StakingInfo';
 import DelegationStepsIntroDialog from '../../source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsIntroDialog';
 import DelegationStepsChooseWalletDialog from '../../source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsChooseWalletDialog';
+import DelegationStepsChooseStakePoolDialog from '../../source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsChooseStakePoolDialog';
 import DelegationStepsNotAvailableDialog from '../../source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsNotAvailableDialog';
 import DelegationStepsConfirmationDialog from '../../source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsConfirmationDialog';
 import DelegationStepsActivationDialog from '../../source/renderer/app/components/staking/delegation-setup-wizard/DelegationStepsActivationDialog';
@@ -25,6 +26,7 @@ import { StakingDelegationCenterStory } from './Staking-DelegationCenter.stories
 import { StakingEpochsStory } from './Staking-Epochs.stories';
 
 import translations from '../../source/renderer/app/i18n/translations';
+import STAKE_POOLS from '../../source/renderer/app/config/stakingStakePools.dummy.json';
 
 const defaultPercentage = 10;
 const defaultStartDateTime = new Date('2019-09-26');
@@ -62,13 +64,22 @@ const WALLETS = [
   },
 ];
 
+const themes = {
+  'Light Blue': 'light-blue',
+  Cardano: 'cardano',
+  'Dark Blue': 'dark-blue',
+};
+
 const locales = {
   English: 'en-US',
   Japanese: 'ja-JP',
 };
+
 // Delegation steps labels are translated outside components and we need to determine correct translations
 const locale = localStorage.getItem('currentLocale') || 'English';
 const translationIndex = locales[locale];
+
+// @TODO - improve locales GET once [DDW-711](https://github.com/input-output-hk/daedalus/pull/1426) is merged
 const DELEGATION_WIZARD_STEPS_LIST = [
   translations[translationIndex]['staking.delegationSetup.steps.step.1.label'],
   translations[translationIndex]['staking.delegationSetup.steps.step.2.label'],
@@ -121,7 +132,7 @@ storiesOf('Staking', module)
     () => (
       <div>
         <StakingCountdown
-          currentLocale="en-US"
+          currentLocale={translationIndex}
           startDateTime={startDateTimeKnob(
             'Decentralization Start DateTime',
             defaultStartDateTime
@@ -196,6 +207,31 @@ storiesOf('Staking', module)
       isSpendingPasswordSet
       onClose={action('onClose')}
       onActivate={action('onActivate')}
+      onBack={action('onBack')}
+    />
+  ))
+
+  .add('DelegationStepsChooseStakePoolDialog', () => (
+    <DelegationStepsChooseStakePoolDialog
+      stepsList={DELEGATION_WIZARD_STEPS_LIST}
+      stakePoolsList={STAKE_POOLS.slice(
+        0,
+        number('Pools', 100, {
+          range: true,
+          min: 37,
+          max: 300,
+          step: 1,
+        })
+      )}
+      stakePoolsDelegatingList={[
+        STAKE_POOLS[0],
+        STAKE_POOLS[13],
+        STAKE_POOLS[36],
+      ]}
+      onOpenExternalLink={() => {}}
+      currentTheme={radios('Theme (Only for tooltip colors)', themes)}
+      onClose={action('onClose')}
+      onContinue={action('onContinue')}
       onBack={action('onBack')}
     />
   ))